+++ src/application.js
@@ -69,8 +69,7 @@ export function startApplication( { baseURL, csrfToken, locale, ...initialState 
   } );
 }
 
-export function destroyApplication()
-{
+export function destroyApplication() {
   app.$destroy();
   app.$options.router.destroy();
   app = null;
