+++ src/routes/admin/index.js
@@ -17,24 +17,8 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 **************************************************************************/
 
-import routeStatic from '@/routes/static'
-import routeIssues from '@/routes/issues'
-import routeAdmin from '@/routes/admin'
-import routeUser from '@/routes/user'
+import routeProjects from '@/routes/admin/projects'
 
-export default function registerRoutes( router, i18n, ajax, store, formatter ) {
-  function routeAll( route ) {
-    routeStatic( route ),
-    routeIssues( route, i18n, ajax, store, formatter ),
-    routeAdmin( route, ajax, store ),
-    routeUser( route )
-  }
-
-  router.register( routeAll );
-
-  if ( process.env.NODE_ENV != 'production' && module.hot != null ) {
-    module.hot.accept( [ '@/routes/static', '@/routes/issues', '@/routes/admin', '@/routes/user' ], () => {
-      router.hotUpdate( routeAll );
-    } );
-  }
+export default function routeAdmin( route, ajax, store ) {
+  routeProjects( route, ajax, store );
 }
