+++ src/routes/user.js
@@ -17,12 +17,10 @@
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 **************************************************************************/
 
-export default function makeUserRoutes() {
-  return function userRoutes( route ) {
-    if ( process.env.TARGET == 'electron' ) {
-      route( 'ClientSettings', '/settings', () => {
-        return Promise.resolve( { form: 'client/ClientSettings', size: 'small' } );
-      } );
-    }
+export default function routeUser( route ) {
+  if ( process.env.TARGET == 'electron' ) {
+    route( 'ClientSettings', '/settings', () => {
+      return Promise.resolve( { form: 'client/ClientSettings', size: 'small' } );
+    } );
   }
 }
